+++ src/main/java/top/theillusivec4/curios/common/CuriosHelper.java
@@ -21,7 +21,10 @@ package top.theillusivec4.curios.common;
 
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Multimap;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
@@ -46,6 +49,7 @@ import org.apache.logging.log4j.util.TriConsumer;
 import top.theillusivec4.curios.api.CuriosApi;
 import top.theillusivec4.curios.api.CuriosCapability;
 import top.theillusivec4.curios.api.SlotContext;
+import top.theillusivec4.curios.api.SlotResult;
 import top.theillusivec4.curios.api.SlotTypePreset;
 import top.theillusivec4.curios.api.type.capability.ICurio;
 import top.theillusivec4.curios.api.type.capability.ICuriosItemHandler;
@@ -94,6 +98,89 @@ public class CuriosHelper implements ICuriosHelper {
     });
   }
 
+  @Override
+  public Optional<SlotResult> findFirstCurio(@Nonnull LivingEntity livingEntity, Item item) {
+    return findFirstCurio(livingEntity, (stack) -> stack.getItem() == item);
+  }
+
+  @Override
+  public Optional<SlotResult> findFirstCurio(@Nonnull LivingEntity livingEntity,
+                                             Predicate<ItemStack> filter) {
+    SlotResult result = getCuriosHandler(livingEntity).map(handler -> {
+      Map<String, ICurioStacksHandler> curios = handler.getCurios();
+
+      for (String id : curios.keySet()) {
+        ICurioStacksHandler stacksHandler = curios.get(id);
+        IDynamicStackHandler stackHandler = stacksHandler.getStacks();
+
+        for (int i = 0; i < stackHandler.getSlots(); i++) {
+          ItemStack stack = stackHandler.getStackInSlot(i);
+
+          if (!stack.isEmpty() && filter.test(stack)) {
+            return new SlotResult(new SlotContext(id, livingEntity, i), stack);
+          }
+        }
+      }
+      return new SlotResult(null, ItemStack.EMPTY);
+    }).orElse(new SlotResult(null, ItemStack.EMPTY));
+    return result.getStack().isEmpty() ? Optional.empty() : Optional.of(result);
+  }
+
+  @Override
+  public List<SlotResult> findCurios(@Nonnull LivingEntity livingEntity, Item item) {
+    return findCurios(livingEntity, (stack) -> stack.getItem() == item);
+  }
+
+  @Override
+  public List<SlotResult> findCurios(@Nonnull LivingEntity livingEntity,
+                                     Predicate<ItemStack> filter) {
+    List<SlotResult> result = new ArrayList<>();
+    getCuriosHandler(livingEntity).ifPresent(handler -> {
+      Map<String, ICurioStacksHandler> curios = handler.getCurios();
+
+      for (String id : curios.keySet()) {
+        ICurioStacksHandler stacksHandler = curios.get(id);
+        IDynamicStackHandler stackHandler = stacksHandler.getStacks();
+
+        for (int i = 0; i < stackHandler.getSlots(); i++) {
+          ItemStack stack = stackHandler.getStackInSlot(i);
+
+          if (!stack.isEmpty() && filter.test(stack)) {
+            result.add(new SlotResult(new SlotContext(id, livingEntity, i), stack));
+          }
+        }
+      }
+    });
+    return result;
+  }
+
+  @Override
+  public List<SlotResult> findCurios(@Nonnull LivingEntity livingEntity, String... identifiers) {
+    List<SlotResult> result = new ArrayList<>();
+    Set<String> ids = Arrays.stream(identifiers).collect(Collectors.toSet());
+    getCuriosHandler(livingEntity).ifPresent(handler -> {
+      Map<String, ICurioStacksHandler> curios = handler.getCurios();
+
+      for (String id : curios.keySet()) {
+
+        if (ids.contains(id)) {
+          ICurioStacksHandler stacksHandler = curios.get(id);
+          IDynamicStackHandler stackHandler = stacksHandler.getStacks();
+
+          for (int i = 0; i < stackHandler.getSlots(); i++) {
+            ItemStack stack = stackHandler.getStackInSlot(i);
+
+            if (!stack.isEmpty()) {
+              result.add(new SlotResult(new SlotContext(id, livingEntity, i), stack));
+            }
+          }
+        }
+      }
+    });
+    return result;
+  }
+
+  @Nonnull
   @Override
   public Optional<ImmutableTriple<String, Integer, ItemStack>> findEquippedCurio(Item item,
                                                                                  @Nonnull